Refactor: styled-component 파일 분리
