+++ src/components/footer/Footer.jsx
@@ -1,0 +1,16 @@
+import * as S from './Style'
+
+function Footer() {
+  return (
+    <>
+      <S.FooterStyle>
+        <div className="max-width">
+          <h2>©Weniv Corp.</h2>
+          <button className="top-button">TOP</button>
+        </div>
+      </S.FooterStyle>
+    </>
+  );
+}
+
+export default Footer;
